Do not use the HDFCreate mode in SD API
In general HDFCreate is just a special case of HDFWrite and there
are no SD interface functions that specifically require HDFCreate
mode (except sd_start). So, the desired behaviour of the HDFCreate
mode could be achieved by calling just sd_start and explicitly
asking for this mode. Latter interactions with the SD interface
could be based on assumption that file is opened with HDFWrite mode.

This commit removes special handling from SD API that allowed the
file-modifying functions to be called only in HDFWrite or HDFCreate
modes. Now only HDFWrite is allowed.