Report error if sd_getattdatainfo found an old-style attribute
If SDgetattdatainfo function is called to get information about an
old-style DFSD API attribute it should inform the user by returning
DFE_NOVGREP. In haskell interface return codes are used only for
reporting failed calls and any extra payload on call status should
be avoided. So, when sd_getattdatainfo finds an old-style attribute
it still returns a regular error code (-1), but it also pushes DFE_NOVGREP
to the error stack. Then the error could be examined by user after
the failed sd_getattdatainfo call to determine whether the failure
was caused by an old-style attribute or by a different reason.