Manually manage memory containing file names passed to HEpush
HEpush expects that pointer to the file name string is a static pointer. So the
error stack just stores this pointer as passed to HEpush and memory should not be
deallocated while the pointer is in error stack. Original implementation used a
temporary C string to store the file name passed to HEpush which was GCed leaving
the HDF error stack pointing to the freed memory. To overcome this problem memory
for storing the file names is allocated and freed manually without relying on GC
and the list of pointers is being kept on the C side to deallocate memory when the
error stack is cleared.