Store NetCDF strings as raw CString values
Using a ByteString to represent NetCDF string type does not play well
with the rest of the NetCDF data types in Haskell wrapper, mainly because
ByteString does not have a Storable instance, so it can not be easily
put into a Storable Vector in a generic way. Thus, for the low-level
NetCDF bindings, using raw CStrings is deemed like an acceptable pragmatic
solution. One major drawback of this approach is that now string memory
allocated by the NetCDF library should be manually freed by calling the
nc_free_string function. Although for the higher-level bindings dedicated
function for working with NetCDF string values could be introduced.