Introduce dedicated singletons representing data types in SomeSDS and SomeNCVariable
Original implementation of SomeSDS used 'HDataType a' as an evidence
parameter which allows to recover the underlying data type (a) used to store
SDS data. But this data type is not unique and can not be directly converted
to HDataType value (for example, knowing that SDS holds Int16 data is not
enough for GHC to deduce that it is an 'HInt16::HDataType Int16' data set).
To allow this a dedicated singleton type 'HDataTypeS (t :: HDataType a)' is
introduced as an evidence in SomeSDS. Thus, using pattern-matching on a
value of HDataTypeS 'HDataType a' can be recovered. Same is applied to the
SomeNCVariable type.