Provide dedicated functions for working with NetCDF4 string variables
When reading a string from a NetCDF4 data set, memory is allocated on
the C-side and has to be manually freed by calling the nc_free_string
function provided by the NetCDF library. Thus, to avoid error-prone
approach with manual memory management, a dedicated set of functions
is introduced. In these functions read ForeignPtr with NetCDF string(s)
is either provided with a finalizer calling nc_free_string or converted
to a ByteString and deallocated right after reading. Thus, calling
nc_free_string manually can be avoided. However, it is still possible
to use general-purpose reading functions for reading NetCDF4 strings, but
in this case memory still should be manually deallocated.

This commit also introduces specialized versions of a few write functions,
which take a single ByteString as an argument.