Report NetCDF deflate level 0 as no compression
Prior to NetCDF 4.9.0 it was possible to have a compressed variable
with deflate level 0 meaning that no actual compression is done. Such
option does not seem to have a lot of practical use, so, to keep behaviour
consistent across different NetCDF versions, the wrapper code is changed
to always report deflate level 0 as no compression.