Rework handling of NetCDF types to allow user-defined types
Original implementation of the NetCDF bindings used a closed set of
atomic types but NetCDF4 allows the user to define their own data types
to be used when storing variables and attributes. Thus to support
user-defined types the closed set of compatible NetCDF/Haskell types is
replaced with an open type family which can be extended by the user.
Currently this initial implementation allows only compound user types.
The user is expected to provide a correct `EquivalentHaskellType` type
instance which properly maps the read NetCDF value onto a Haskell one.