Optimize the TTernary type family to avoid exponential complexity
Since for most non-trivial cases current GHC versions try to first fully
reduce the condition and _both_ branches of an If before picking,
the original implementation of TTernary quickly becomes prohibitively
slow and memory-hungry. To avoid this If-based logic is replaced by a
dedicated helper type family which ensures that we reduce only what
we actually need.